Added value clipping for last layer to fix the numerical instability in the ctc loss calculation that caused the training to fail after a number of epochs.

+++ data.py
@@ -25,11 +25,11 @@ def _parse_function(proto):
     
     #reference = tf.sparse.to_dense(reference)
     
-    # Convert chunk and reference_length as they are
+    # Convert chunk and reference_length
     chunk = tf.cast(parsed_features['chunk'], tf.float32)
     # For some reason the target_lengths in the train dataset are offset by 7!
     length = tf.minimum(parsed_features['reference_length'] + 7, 500)
-    reference_length = tf.cast(length, tf.int64)
+    reference_length = tf.cast(length, tf.int32)
     
     return chunk, reference, reference_length
